+++ lcheapo/tests/test_lcheapo.py
@@ -7,12 +7,13 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 from future.builtins import *  # NOQA @UnusedWildImport
 
-import os
+from os import system
 import unittest
 import filecmp
 import inspect
 import difflib
 import json
+from pathlib import Path
 
 
 class TestLCHEAPOMethods(unittest.TestCase):
@@ -20,10 +21,9 @@ class TestLCHEAPOMethods(unittest.TestCase):
     Test suite for nordic io operations.
     """
     def setUp(self):
-        self.path = os.path.dirname(os.path.abspath(inspect.getfile(
-            inspect.currentframe())))
-        self.testing_path = os.path.join(self.path, "data")
-        self.exec_path = os.path.split(self.path)[0]
+        self.path = Path(inspect.getfile(
+            inspect.currentframe())).resolve().parent
+        self.testing_path = self.path / "data"
 
     def assertProcessStepsFilesEqual(self, first, second, msg=None):
         with open(first, "r") as fp:
@@ -72,16 +72,12 @@ class TestLCHEAPOMethods(unittest.TestCase):
         Test lcdump outputs.
         """
         # WRITEOUT OF DATA HEADERS
-        # cmd = os.path.join(self.exec_path, 'lcdump.py') + ' ' +\
-        cmd = 'lcdump' + ' ' +\
-            os.path.join(self.testing_path, 'BUGGY.raw.lch') +\
-            ' 5000 100  > temp_test.out'
-        os.system(cmd)
+        cmd = f'lcdump {Path(self.testing_path) / "BUGGY.raw.lch"} 5000 100  > temp_test.out'
+        system(cmd)
         self.assertTextFilesEqual(
             'temp_test.out',
-            os.path.join(self.testing_path, 'BUGGY_lcdump_5000_100.txt')
-            )
-        os.remove('temp_test.out')
+            Path(self.testing_path) / 'BUGGY_lcdump_5000_100.txt')
+        Path('temp_test.out').unlink()
 
         # WRITEOUT OF FILE HEADER
 
@@ -92,40 +88,38 @@ class TestLCHEAPOMethods(unittest.TestCase):
         Test lcfix on a typical (buggy) file
         """
         # Run the code
-        # cmd = os.path.join(self.exec_path, 'lcfix.py') + \
         cmd = f'lcfix -d {self.path} -i data BUGGY.raw.lch > temp'
-        os.system(cmd)
-        os.remove('temp')
-        # print(os.listdir('.'))
+        system(cmd)
+        Path('temp').unlink()
 
         # Check that the appropriate files were created
-        assert not os.path.exists('BUGGY.fix.timetears.txt')
+        assert not Path('BUGGY.fix.timetears.txt').exists()
 
         # Compare binary files (fix.lch)
         outfname = 'BUGGY.fix.lch'
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
         self.assertBinFilesEqual(
             outfname,
-            os.path.join(self.testing_path, outfname))
-        os.remove(outfname)
+            Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
 
         # Compare text files (fix.txt)
         outfname = 'BUGGY.fix.txt'
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
         self.assertTextFilesEqual(
             outfname,
-            os.path.join(self.testing_path, outfname))
-        os.remove(outfname)
+            Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
 
         # Compare text files (process-steps.json)
         outfname = 'process-steps.json'
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
         new_outfname = 'BUGGY.' + outfname
-        os.rename(outfname, new_outfname)
+        Path(outfname).rename(new_outfname)
         self.assertProcessStepsFilesEqual(
             new_outfname,
-            os.path.join(self.testing_path, new_outfname))
-        os.remove(new_outfname)
+            Path(self.testing_path) / new_outfname)
+        Path(new_outfname).unlink()
 
     def test_lcfix_bad(self):
         """
@@ -133,37 +127,37 @@ class TestLCHEAPOMethods(unittest.TestCase):
         """
         # Run the code
         cmd = f'lcfix -d {self.path} -i data BAD.bad.lch > temp'
-        os.system(cmd)
-        os.remove('temp')
+        system(cmd)
+        Path('temp').unlink()
 
         # Confirm that no lch file was created
-        assert not os.path.exists('BAD.fix.lch')
+        assert not Path('BAD.fix.lch').exists()
 
         # Compare text files (fix.txt)
         outfname = 'BAD.fix.txt'
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
         self.assertTextFilesEqual(
             outfname,
-            os.path.join(self.testing_path, outfname))
-        os.remove(outfname)
+            Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
 
         # Compare text files (fix.timetears.txt)
         outfname = 'BAD.fix.timetears.txt'
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
         self.assertTextFilesEqual(
             outfname,
-            os.path.join(self.testing_path, outfname))
-        os.remove(outfname)
+            Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
 
         # Compare process-steps files
         outfname = 'process-steps.json'
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
         new_outfname = 'BAD.' + outfname
-        os.rename(outfname, new_outfname)
+        Path(outfname).rename(new_outfname)
         self.assertProcessStepsFilesEqual(
             new_outfname,
-            os.path.join(self.testing_path, new_outfname))
-        os.remove(new_outfname)
+            Path(self.testing_path) / new_outfname)
+        Path(new_outfname).unlink()
 
     def test_lccut(self):
         """
@@ -171,17 +165,17 @@ class TestLCHEAPOMethods(unittest.TestCase):
         """
         # Run the code
         cmd = f'lccut -d {self.path} -i data BUGGY.fix.lch --start 5000 --end 5099 > temp'
-        os.system(cmd)
-        os.remove('temp')
-        os.remove('process-steps.json')
+        system(cmd)
+        Path('temp').unlink()
+        Path('process-steps.json').unlink()
 
         # Compare binary files (fix.timetears.txt)
         outfname = 'BUGGY.fix_5000_5099.lch'
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
         self.assertBinFilesEqual(
             outfname,
-            os.path.join(self.testing_path, outfname))
-        os.remove(outfname)
+            Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
 
     def test_lcinfo(self):
         """
@@ -189,33 +183,44 @@ class TestLCHEAPOMethods(unittest.TestCase):
         """
         # Run the code
         cmd = f'lcinfo -d {self.path} -i data BUGGY.fix.lch > temp'
-        os.system(cmd)
+        system(cmd)
 
         # Compare text files
         self.assertTextFilesEqual(
             'temp',
-            os.path.join(self.testing_path, 'BUGGY.info.txt')
-            )
-        os.remove('temp')
+            Path(self.testing_path) / 'BUGGY.info.txt')
+        Path('temp').unlink()
 
     def test_lcheader(self):
         """
         Test lcheader
         """
-        # Run the code
+        # Run the code without any questions
         cmd = f'lcheader --no_questions'
-        os.system(cmd)
+        system(cmd)
 
         outfname = 'generic.header.lch'
         # Check that the appropriate file was created
-        assert os.path.exists(outfname)
+        assert Path(outfname).exists()
 
         # Compare output binary file (fix.lch)
-        self.assertBinFilesEqual(
-            outfname,
-            os.path.join(self.testing_path, outfname))
-        os.remove(outfname)
-        os.remove('process-steps.json')
+        self.assertBinFilesEqual(outfname,
+                                 Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
+
+        # Run the code with all specified
+        cmd = f'lcheader -d MOMARL_SPOBS2_04_LSVEL -s 62.5 -c 4 -w 2019-02-04T04:53:30.024 -e 2019-06-14T04:04:29 -o LSVEL.header.lch > LSVEL.txt'
+        system(cmd)
+        outfname = 'LSVEL.header.lch'
+        assert Path(outfname).exists()
+        self.assertBinFilesEqual(outfname,
+                                 Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
+        outfname = 'LSVEL.txt'
+        self.assertTextFilesEqual(outfname,
+                                  Path(self.testing_path) / outfname)
+        Path(outfname).unlink()
+        Path('process-steps.json').unlink()
 
 
 def suite():